Changed the way magnetometer calibration works.
Before only the maximum value in all directions was used to fit an ellipsoid and scale it onto a unit sphere.
Now also rotation of an ellipsoid are considered allowing the compensation for hard and soft iron errors, scaling erros
and non-orthogonality errors. The vector of raw magnetomer values is therefore processed in the following way:
	- substract the offset vector (center of the ellipsoid)
	- multiply a transformation matrix (from left side). The transformation matrix will scale the values, so that the vector lies close to the unit sphere
	   and apply the correct coordinate tranformation (mainly small rotation) to correct for soft iron errors.